refactor(lib::songtag): rename "new_netease" to "netease_v2"
